Drop rsyslog dependency
cloud-init currently requires rsyslog for its /etc/rsyslog.d directory,
but it does not require rsyslog to be installed or running to function.
The current packaging guidelines now allow for multiple packages to own
directories when they are not required for proper functionality, so we
can now drop this dependency.

https://bugzilla.redhat.com/show_bug.cgi?id=986511